Do not update API payment unless GOV.UK success event is recorded
There have been some problems with the received_at time on payments
being set as too early. This is may be due to the success event
either not existing or not being in its final state at the point
at which the payment is confirmed as succesful, so instead we should
update the received_at some time afterwards.