Avoid sending emails twice for successful payments
As we update the payments asynchronously, every payment will typically
be processed twice, once synchronously (with no update) and once
asynchronously (with an update). As in order to ensure that an email
confirmation is received (and promptly) both of these cases result in
an email being sent, it is necessary to add a way to check if it has
been sent already. This is done buy updating that email field on the
payment when the email is sent, and checking it before sending it again.