Get capture time from new settlement_summary field
This is a new JSON block added to the payment response containing
accurate capture information. It includes the time the capture was
submitted, and the date that it was actually captured. If the
captured_date is later than the submit time, the received_at
attribute of the payment is set to match the captured_date
in order to match the settlement batches we receive from worldpay.