Send confirmation email only when capture date available
Before this, we were sending a confirmation email when capturing
the payment (implicitly with non-delayed payment or explicitly
when calling the GOV.UK /capture on delayed payments).
However the GOV.UK operation is async and Pay captures the payment
at a later time, even after days.
It's quite unlikely but Pay could encounter difficulties when
capturing the payment and in these cases the payment status
would change from success to failed.

For this reason, it's safer to send the confirmation email when
the capture date is available so that we know for sure that
the payment was actually captured.