Remove payment type selector from support form
This commit specifically removes the selector from the 'contact us
about a payment you've sent' form - the selector on the 'contact us
about a payment you're trying to make' was removed in an earlier
commit.

I've removed references to payment type from what we add to these
tickets rather than defaulting it to 'debit card'. I did this to avoid
someone reading the ticket being misled (since the sender could, as an
outside chance, be contacting us about a bank transfer in error).

(JIRA: MTP-1626)