Remove payment ID from GOV.UK Pay Exceptions message
We can get quite a bit of these exceptions at times and Sentry should
really be able to group them as they're all the same class of problem.

As the Payment ID is in the error message these are all considerate as
separate issues.

I've removed the payment ID from the error message and moved this potentially
useful information in the Sentry Context so that it's still available
in the issue but it doesn't prevent issues grouping.

NOTE: These errors are sent to Sentry via the `logger.error()`, I'm not
using Sentry contexts explicitly but Sentry integration is clever enough
to add the values in the dictionary passed to the logger in the Sentry
issue.

See Sentry contexts: https://develop.sentry.dev/sdk/event-payloads/contexts/

Ticket: https://dsdmoj.atlassian.net/browse/MTP-1794